deploy-tool-nithin/requirements.txt
click
boto3
GitPython
PyYAML
